lokasewa-evaluator/requirements.txt
# Core Dependencies
langgraph>=0.2.45
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# AI Model APIs (for fallback and direct API calls)
google-generativeai>=0.8.0
openai>=1.0.0

# File Processing
pdf2image>=1.16.0
Pillow>=10.0.0

# HTTP and Async
aiohttp>=3.9.0
httpx>=0.25.0

# Utility
python-multipart>=0.0.6
asyncio

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0